requirements.txt
Babel==2.12.1
beautifulsoup4==4.11.2
blessed==1.20
coverage==7.4.0
coverage[toml]
croniter
cryptography==43.0.1
Deprecated~=1.2.13
discord.py==2.4
django==5.0.9
django-admin-interface==0.29.2
django-cors-headers==4.4.0
django-coverage-plugin==3.1.0
django-debug-toolbar==4.4.6
django-environ==0.11.2
django-extensions==3.2.3
django-filter==24.3
django-ical==1.9.2
django-q2==1.6.2
django-request==1.6.3
djangorestframework==3.15.2
drf-spectacular==0.27.2
factory-boy==3.3.1
hiredis==3.0.0
niquests==3.9.1
psutil==6.1.0
psycopg2==2.9.9
pymemcache==4.0.0
pymongo==4.10.1
python-dateutil==2.8.2
python-telegram-bot==13.15
redis==5.1.1
requests==2.32.3
